refactor(loss_hub): colocate algorithm prepare and loss hooks
Make Flow-GRPO the explicit actor fallback and keep NFT-specific preparation and loss behavior together without resolver helpers.
